# Remove excessive floating-point divides