--- a/src/lstm/networkio.cpp
+++ b/src/lstm/networkio.cpp
@@ -224,6 +224,7 @@ void NetworkIO::Copy2DImage(int batch, Image pix, float black, float contrast, T
   int target_width = stride_map_.Size(FD_WIDTH);
   int num_features = NumFeatures();
   bool color = num_features == 3;
+  float inv_contrast = 1.0/contrast;
   if (width > target_width) {
     width = target_width;
   }
@@ -236,11 +237,11 @@ void NetworkIO::Copy2DImage(int batch, Image pix, float black, float contrast, T
           int f = 0;
           for (int c = COLOR_RED; c <= COLOR_BLUE; ++c) {
             int pixel = GET_DATA_BYTE(line + x, c);
-            SetPixel(t, f++, pixel, black, contrast);
+            SetPixel(t, f++, pixel, black, inv_contrast);
           }
         } else {
           int pixel = GET_DATA_BYTE(line, x);
-          SetPixel(t, 0, pixel, black, contrast);
+          SetPixel(t, 0, pixel, black, inv_contrast);
         }
       }
     }
@@ -264,6 +265,7 @@ void NetworkIO::Copy1DGreyImage(int batch, Image pix, float black, float contras
   index.AddOffset(batch, FD_BATCH);
   int t = index.t();
   int target_width = stride_map_.Size(FD_WIDTH);
+  float inv_contrast = 1.0/contrast;
   if (width > target_width) {
     width = target_width;
   }
@@ -272,7 +274,7 @@ void NetworkIO::Copy1DGreyImage(int batch, Image pix, float black, float contras
     for (int y = 0; y < height; ++y) {
       uint32_t *line = pixGetData(pix) + wpl * y;
       int pixel = GET_DATA_BYTE(line, x);
-      SetPixel(t, y, pixel, black, contrast);
+      SetPixel(t, y, pixel, black, inv_contrast);
     }
   }
   for (; x < target_width; ++x) {
@@ -287,8 +289,9 @@ void NetworkIO::Copy1DGreyImage(int batch, Image pix, float black, float contras
 // pixel: the value of the pixel from the image (in one channel)
 // black: the pixel value to map to the lowest of the range of *this
 // contrast: the range of pixel values to stretch to half the range of *this.
-void NetworkIO::SetPixel(int t, int f, int pixel, float black, float contrast) {
-  float float_pixel = (pixel - black) / contrast - 1.0f;
+// inv_contrast: one over the contrast, to save a divide
+void NetworkIO::SetPixel(int t, int f, int pixel, float black, float inv_contrast) {
+  float float_pixel = (pixel - black) * inv_contrast - 1.0f;
   if (int_mode_) {
     i_[t][f] = ClipToRange<int>(IntCastRounded((INT8_MAX + 1) * float_pixel), -INT8_MAX, INT8_MAX);
   } else {