--- a/src/lstm/networkio.cpp
+++ b/src/lstm/networkio.cpp
@@ -224,7 +224,7 @@ void NetworkIO::Copy2DImage(int batch, Image pix, float black, float contrast, T
   int target_width = stride_map_.Size(FD_WIDTH);
   int num_features = NumFeatures();
   bool color = num_features == 3;
-  float inv_contrast = 1.0/contrast;
+  float inv_contrast = 1 / contrast;
   if (width > target_width) {
     width = target_width;
   }
@@ -265,7 +265,7 @@ void NetworkIO::Copy1DGreyImage(int batch, Image pix, float black, float contras
   index.AddOffset(batch, FD_BATCH);
   int t = index.t();
   int target_width = stride_map_.Size(FD_WIDTH);
-  float inv_contrast = 1.0/contrast;
+  float inv_contrast = 1 / contrast;
   if (width > target_width) {
     width = target_width;
   }