@stweil 2024-09-03
Do you have timing values from your tests?