@heshpdx 2024-09-03
> Do you have timing values from your tests?

It will be CPU specific, but I see +10% on my Ampere Altra.