@stweil 2024-09-03
> It will be CPU specific, but I see +10% on my Ampere Altra.

That's a very significant improvement!