@heshpdx 2024-09-03
If there are standard tests that you run, please do share the results.

I was using ` -l deu --tessdata-dir ./tessdata_orig --oem 0` and `-l Arabic --tessdata-dir ./tessdata_fast`. The `-l rus --tessdata-dir ./tessdata_orig --oem 2` did not show much improvement.