@heshpdx 2024-09-23
After some wrangling, I was able to get the unit tests running on my machine. Here is a rollup of the tests which run longer than 1ms total. I basically culled this out using `grep total *_test.log | grep ran | grep -v 0\  | tr '(' ' ' | sort`. I hope this is sufficient!

test | before | after | speedup
-- | -- | -- | --
apiexample_test | 454 | 448 | 1.013
applybox_test | 809 | 799 | 1.013
baseapi_test | 4582 | 4485 | 1.022
baseapi_thread_test | 424 | 415 | 1.022
commandlineflags_test | 39 | 39 | 1.000
imagedata_test | 281 | 272 | 1.033
intfeaturemap_test | 1021 | 1012 | 1.009
intsimdmatrix_test | 825 | 826 | 0.999
lang_model_test | 7829 | 7782 | 1.006
layout_test | 7232 | 6665 | 1.085
ligature_table_test | 4 | 4 | 1.000
loadlang_test | 141 | 138 | 1.022
lstm_recode_test | 65028 | 64371 | 1.010
lstm_squashed_test | 28698 | 29057 | 0.988
lstm_test | 252497 | 246625 | 1.024
normstrngs_test | 14 | 14 | 1.000
pagesegmode_test | 311 | 296 | 1.051
pango_font_info_test | 13 | 13 | 1.000
paragraphs_test | 54 | 54 | 1.000
progress_test | 6 | 5 | 1.200
qrsequence_test | 81 | 80 | 1.013
recodebeam_test | 1128 | 1108 | 1.018
stringrenderer_test | 177 | 177 | 1.000
tablefind_test | 73 | 72 | 1.014
tatweel_test | 16 | 17 | 0.941
textlineprojection_test | 2269 | 2108 | 1.076
unicharcompress_test | 214 | 208 | 1.029




```
 
```