@zdenop 2024-09-24
> With the latest changes, I get +25% on this cmdline.

@heshpdx : Is it with or without simd usage?