@heshpdx 2024-09-24
Good question. This was using the generic path, without intrinsics.