@stweil 2024-09-24
If `tesseract -v` reports "Found NEON", it will use SIMD instructions for NEON.